Generate attribution HTML after dereferencing the license.
ccffext does not have access to the browser to retrieve the license
and parse it for RDFa. We rely on some information (title, in
particular) from the license document for generating attribution
HTML. Moving this into the getLicenseDetails callback ensures that we
only attempt to use that information after it's available.