Issue-36 Handle redirects when the target contains valid assertions.
When the license resource is redirected (ie, over https by the HTTPS
Everywhere add-on), store the original URL. If the target contains
assertions which match the original URL, honor them. Previously the
assertions where cached in the incorrect "bucket" (under the
redirected, https, URI, instead of the initial http URI).